skills/image-generator/pyproject.toml
[project]
name = "image-generator-skill"
version = "0.1.0"
description = "Generate batches of photorealistic image variations using Microsoft Foundry (MAI-Image-2 and gpt-image-1.5)."
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.3",
    "httpx>=0.28.1",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.4",
]

[project.scripts]
generate-images = "image_generator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_generator"]
